♻️ Typing improvements
